# Faster and less memory hungry version of gdistances