#495 Here is the proposed code. I have moved the definitions to the end of the file as they use custom BFS implemenation.

Here is the performance testing code I used:
```
g = smallgraph(:tutte)
println("tutte 3")
@time a = gdistances(g,3);

srand(1)
for i in 0:5
    g = watts_strogatz(20000, 100, 0.5)
    vl = sample(1:100, 5i, replace=false)
    println("Watts-Strogatz ", vl)
    @time a = gdistances(g,vl);
end
```

It seems to use less time and memory than the current implementation. I have also updated tests for a corner case of emply `source` collection.