@sbromberger 2016-10-31
@bkamins - thanks. Could you post benchmarks for this in this thread? Specifically, try both sparse and dense graphs, with single- and multi-source. That way we'll have it in the record here and it'll be another verification.

@jpfairbanks take a look at this - even if we revert the regression in https://github.com/JuliaGraphs/LightGraphs.jl/commit/c571fdfdf9bee2a01c15dc989cdebfe750a9211f, @bkamins' code is 5x faster than using the generic bfs traversal.