@bkamins 2016-10-31
Here are the benchmark results:
```
tutte graph 50% source
old:  0.007651 seconds (1.89 k allocations: 81.509 KB)
new:  0.000004 seconds (2 allocations: 896 bytes)

tutte graph single source
old:  0.000067 seconds (145 allocations: 5.516 KB)
new:  0.000003 seconds (2 allocations: 896 bytes)

large sparse graph 50% source
old:  4.211344 seconds (40.98 M allocations: 794.589 MB, 3.60% gc time)
new:  0.399337 seconds (4 allocations: 15.259 MB, 0.77% gc time)

large sparse graph single source
old:  4.230289 seconds (40.98 M allocations: 790.589 MB, 3.59% gc time)
new:  0.386659 seconds (4 allocations: 15.259 MB, 0.65% gc time)

medium dense graph 50% source
old: 30.604378 seconds (447.57 M allocations: 8.162 GB, 5.25% gc time)
new:  0.165119 seconds (4 allocations: 1.526 MB)

medium dense graph single source
old: 30.385432 seconds (447.57 M allocations: 8.160 GB, 5.27% gc time)
new:  0.140546 seconds (4 allocations: 1.526 MB)

small very dense graph 50% source
old:  3.042858 seconds (44.49 M allocations: 831.601 MB, 5.47% gc time)
new:  0.015054 seconds (4 allocations: 78.281 KB)

small very dense graph single source
old:  3.039986 seconds (44.49 M allocations: 831.601 MB, 5.47% gc time)
new:  0.014593 seconds (4 allocations: 78.281 KB)
```

and the code (old is current master, new is the proposed implementation):
```
using LightGraphs
using Distributions

function new_gdistances!(g::SimpleGraph, source, dists)
    n = nv(g)
    fill!(dists, -1)
    queue = fill(-1, n)
    for i in 1:length(source)
        queue[i] = source[i]
        dists[source[i]] = 0
    end
    head = 1
    tail = length(source)
    while head <= tail
        current = queue[head]
        distance = dists[current] + 1
        head += 1
        for j in fadj(g, current)
            if dists[j] == -1
                dists[j] = distance
                tail += 1
                queue[tail] = j
            end
        end
    end
    dists
end

function new_gdistances(g::SimpleGraph, source)
    new_gdistances!(g, source, Vector{Int}(nv(g)))
end

function run_test(g, multi, caption)
    vl = sample(1:nv(g),
                multi ? div(nv(g), 2) : 1,
                replace = false)
    println("\n", caption, multi ? " 50% source":" single source")
    print("old:")
    @time a = gdistances(g,vl);
    print("new:")
    @time b = new_gdistances(g,vl);
    @assert a == b
end

srand(1)
gs = [(smallgraph(:tutte), "tutte graph"),
      (watts_strogatz(10^6, 10, 1), "large sparse graph"),
      (watts_strogatz(10^5, 1000, 0), "medium dense graph"),
      (watts_strogatz(5000, 2000, 0.5), "small very dense graph")]
@assert gdistances(gs[1][1], 3) == new_gdistances(gs[1][1], 3)

for (g, c) in gs, m in [true, false]
    run_test(g, m, c)
end
```