@sbromberger 2016-10-31
Beautiful. Thank you for doing that extra work. Those numbers are compelling!