@jpfairbanks 2016-11-01
Can we add these benchmark cases to the tests? Up to 10^4 vertices for speed. Including the old definition in the test source code so that we can check for correctness? And add the full benchmarks to a file `/bench/gdistances.jl`