@sbromberger 2016-11-01
> Can we add these benchmark cases to the tests? Up to 10^4 vertices for speed. Including the old definition in the test source code so that we can check for correctness? And add the full benchmarks to a file /bench/gdistances.jl

Wait - no. I don't like this idea. We should do this when we get #45 implemented but there's no point in doing that now, especially adding the old code back in.

The old code was slower - originally by a factor of 5, but then we had a very bad regression in https://github.com/JuliaGraphs/LightGraphs.jl/commit/c571fdfdf9bee2a01c15dc989cdebfe750a9211f that caused a massive performance hit. This code is the fastest implementation yet.