@jpfairbanks 2016-11-01
The code in master doesn't look like it would be that much slower than what is written here. Can we do some probing into why the code in master is so much slower? We could probably fix a performance bug that makes all the traversals faster. I'd rather fix a bug in the existing BFS traversal code, than use a custom BFS implementation.