@sbromberger 2016-11-01
The code in master overallocates since https://github.com/JuliaGraphs/LightGraphs.jl/commit/c571fdfdf9bee2a01c15dc989cdebfe750a9211f effectively got rid of `gdistances!`.