@jpfairbanks 2016-11-01
hmm... GdistanceVisitor never gets a definition for `examine_neighbor!`