@jpfairbanks 2016-11-01
> The  code in master overallocates since c571fdf effectively got rid of gdistances!.

Which array is overallocated?