@jpfairbanks 2016-11-01
Here is a basic profiling result based on @bkamins' benchmarkk
```julia
julia> g = watts_strogatz(10000, 2000, 0.5)
{10000, 10000000} undirected graph

julia> @profile gdistances(g, 3); Base.Profile.print()
4533 ./event.jl:68; (::Base.REPL.##3#4{Base.REPL.REPLBackend})()
 4533 ./REPL.jl:95; macro expansion
  4533 ./REPL.jl:64; eval_user_input(::Any, ::Base.REPL.REPLBackend)
   4533 ./boot.jl:234; eval(::Module, ::Any)
    4533 ./<missing>:?; anonymous
     4533 ./profile.jl:16; macro expansion;
      4533 ...ia/v0.5/LightGraphs/src/traversals/bfs.jl:113; gdistances(::LightGraphs.Graph, ::Int64)
       4533 ...ia/v0.5/LightGraphs/src/traversals/bfs.jl:99; gdistances!(::LightGraphs.Graph, ::Int64, ::Array{Int64,1})
        4533 ./<missing>:0; (::LightGraphs.#kw##traverse_graph!)(::Array{Any,1}, ::LightGr...
         4533 ...a/v0.5/LightGraphs/src/traversals/bfs.jl:73; #traverse_graph!#15(::Array{Int64,1}, ::LightGraphs.DummyEdg...
          2048 .../v0.5/LightGraphs/src/traversals/bfs.jl:40; breadth_first_visit_impl!(::LightGraphs.Graph, ::Array{Int64...
           14 ./array.jl:0; done(::Array{Int64,1}, ::Int64)
           4  ./array.jl:381; done(::Array{Int64,1}, ::Int64)
           56 ./array.jl:380; next(::Array{Int64,1}, ::Int64)
           7  ./bool.jl:0; !(::Bool)
           9  ./bool.jl:16; !(::Bool)
           1  ...mes/.julia/v0.5/LightGraphs/src/core.jl:243; out_neighbors(::LightGraphs.Graph, ::Int64)
          404  .../v0.5/LightGraphs/src/traversals/bfs.jl:41; breadth_first_visit_impl!(::LightGraphs.Graph, ::Array{Int64...
           4  ./abstractarray.jl:0; get(::Array{Int64,1}, ::Int64, ::Int64)
           26 ./abstractarray.jl:912; get(::Array{Int64,1}, ::Int64, ::Int64)
          1169 .../v0.5/LightGraphs/src/traversals/bfs.jl:42; breadth_first_visit_impl!(::LightGraphs.Graph, ::Array{Int64...
           8 ./essentials.jl:0; convert(::Type{Int64}, ::Int64)
          547  .../v0.5/LightGraphs/src/traversals/bfs.jl:44; breadth_first_visit_impl!(::LightGraphs.Graph, ::Array{Int64...
          322  .../v0.5/LightGraphs/src/traversals/bfs.jl:46; breadth_first_visit_impl!(::LightGraphs.Graph, ::Array{Int64...
           1  ./promotion.jl:0; ==(::Int64, ::Int64)
           10 ./promotion.jl:269; ==(::Int64, ::Int64)

```


This appears to be the offender:  `1169 .../v0.5/LightGraphs/src/traversals/bfs.jl:42; breadth_first_visit_impl!(::LightGraphs.Graph, ::Array{Int64...`

And Line 42 in the source is `v_edge = Edge(u,v)` 

Which is inconsistent with my conception of the Pair{Int,Int} type being put on the stack instead of the heap.