@jpfairbanks 2016-11-01
The only significant difference between the two codes, should be the `push!` on line 49 of bfs.jl. The implementation in this PR uses a fixed size queue. Which we could do in the Visitor methods currently implemented.